Nuestros cambios

+++ app/src/main/java/com/example/marketpets/Models/Comida.kt
@@ -4,7 +4,7 @@ data class Comida (
     val nombre: String? = null,
     val descripcion: String? = null,
     val marca: String? = null,
-    val precio: Double? = null,
+    val precio: Int? = null,
     val tamano: String? = null,
     val disponibilidad: Int? = null,
     val imagen: String?
